Possible Release 12.9.8.3 fix to earlier commit(check for amdkmafd remove status.)

+++ trunk/ display-drivers-uninstaller/WindowsApplication1/Form1.vb
@@ -6499,21 +6499,24 @@ Public Class CleanupEngine
                     If regkey IsNot Nothing Then
                         For i As Integer = 0 To driverfiles.Length - 1
                             If Not checkvariables.isnullorwhitespace(driverfiles(i)) Then
-                                For Each child As String In regkey.GetValueNames()
-                                    If checkvariables.isnullorwhitespace(child) = False Then
-                                        If child.ToLower.Contains(driverfiles(i).ToLower) Then
-                                            Try
-                                                regkey.DeleteValue(child)
-                                            Catch ex As Exception
-                                                f.log(ex.Message & " @Pnplockdownfiles")
-                                            End Try
+                                If Not (donotremoveamdhdaudiobusfiles AndAlso driverfiles(i).ToLower.Contains("amdkmafd")) Then
+                                    For Each child As String In regkey.GetValueNames()
+                                        If checkvariables.isnullorwhitespace(child) = False Then
+                                            If child.ToLower.Contains(driverfiles(i).ToLower) Then
+                                                Try
+                                                    regkey.DeleteValue(child)
+                                                Catch ex As Exception
+                                                    f.log(ex.Message & " @Pnplockdownfiles")
+                                                End Try
+                                            End If
                                         End If
-                                    End If
-                                Next
+                                    Next
+                                End If
                             End If
                         Next
                     End If
                 End If
+
             End If
 
         Catch ex As Exception
